% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{dl_basin}
\alias{dl_basin}
\title{Download water quality data for a basin}
\usage{
dl_basin(basin)
}
\arguments{
\item{basin}{the name of a basin.
An easy way to get a list of basins is to use the \code{\link{pt_basins}}
function}
}
\value{
a data.frame of all the water quality monitoring data from that basin.
}
\description{
Download water quality data for a basin
}
