% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dl_site}
\alias{dl_site}
\title{Download water quality data for a site}
\usage{
dl_site(site)
}
\arguments{
\item{site}{site number. See \code{\link{wq_sites}}.}
}
\value{
a data.frame of water quality data for the site. See \code{\link{wq_param_desc}}
}
\description{
Download water quality data for a site
}
\examples{
\dontrun{
dl_site("YT09FC0002")
}
}
