% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{dl_sites}
\alias{dl_sites}
\title{Download water quality data for a site or multiple sites}
\usage{
dl_sites(sites)
}
\arguments{
\item{sites}{site numbers. See \code{\link{wq_sites}}.}
}
\value{
a data.frame of water quality data for the sites.
See \code{\link{wq_param_desc}}
}
\description{
Download water quality data for a site or multiple sites
}
\examples{
\dontrun{
dl_sites(c("NW10OB0003", "NB01AJ0008"))
}
}
