% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{pt_basins}
\alias{pt_basins}
\title{Get a list of basins for a Province or Territory.}
\usage{
pt_basins(prov_terr = c("AB", "BC", "MB", "NB", "NL", "NS", "NT", "NU", "ON",
  "PE", "QC", "SK", "US", "YT"))
}
\arguments{
\item{prov_terr}{one or more Province/Territory abbreviations}
}
\value{
a character vector of basin names
}
\description{
Get a list of basins for a Province or Territory.
}
\examples{
\dontrun{
  pt_basins(c("BC", "AB"))
}
}
