% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{wq_basins_pt}
\alias{wq_basins_pt}
\title{Get a list of basins for a Province or Territory.}
\usage{
wq_basins_pt(prov_terr = c("AB", "BC", "MB", "NB", "NL", "NS", "NT", "NU",
  "ON", "PE", "QC", "SK", "US", "YT"))
}
\arguments{
\item{prov_terr}{one or more Province/Territory abbreviations}
}
\value{
a character vector of basin names
}
\description{
Get a list of basins for a Province or Territory.
}
\examples{
\dontrun{
  wq_basins_pt(c("BC", "AB"))
}
}
