% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{wq_data_desc}
\alias{wq_data_desc}
\title{Get a table of water quality parameter descriptions}
\usage{
wq_data_desc()
}
\value{
a data.frame of parameter descriptions
}
\description{
Get a table of water quality parameter descriptions
}
