% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{wq_sites}
\alias{wq_sites}
\title{Get a table of water quality monitoring sites}
\usage{
wq_sites()
}
\value{
a data.frame of monitoring sites
}
\description{
Get a table of water quality monitoring sites
}
