% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canwqdata-package.r
\docType{package}
\name{canwqdata}
\alias{canwqdata}
\alias{canwqdata-package}
\title{Get Canadian Long-Term Water Quality Monitoring Data}
\description{
All data is freely available under the
\href{http://open.canada.ca/en/open-government-licence-canada}{Open Government License Canada}
from \href{http://data.ec.gc.ca/data/substances/monitor/national-long-term-water-quality-monitoring-data/}{here}.
}
